UnitsSinceLastOperation, WarningBoundary

+++ src/PrizmMainProject/Common/Constants.cs
@@ -25,7 +25,7 @@ namespace Prizm.Main.Common
         /// <summary>
         /// period in percents for remaining about execution of control operation
         /// </summary>
-        public const double PercentForInspectionOperation = 0.90;
+        public const float PercentForInspectionOperation = 0.90f;
 
         /// <summary>
         /// Warning should be shown this amount of days before expiration
